requirements.txt
google-adk==0.2.0
yfinance
psutil==5.9.5
